Add types for undocumented methods
Some methods in `VLCPlayer.js` exist and were functional but did not
have typings so TypeScript complained when you tried to use them.
These have been added to the declarations and the declarations.

The declarations have also been cleaned up to use a single base class
for methods shared across `VLCPlayer` and `VlCPlayerView`.